Fix bug where monitored collection would hang if http filter short-circuited the remote server.
The TrafficMonitorFilter had been patterned after BMP's HarCaptureFilter,
but that filter doesn't capture responses that are returned by any HttpFilters
an outgoing request goes through. The TrafficMonitorFilter is hereby revised
to capture that response and notify the listener about a request/response
pair. This may or may not have to do with the hanging, but the change appears
to resolve the issue.