use ChromeOptions.setAcceptInsecureCerts(true) to accommodate ChromeDriver 79 changes
See https://chromedriver.storage.googleapis.com/79.0.3945.16/notes.txt
and https://bugs.chromium.org/p/chromedriver/issues/detail?id=3148

Previously, ChromeDriver always passed the --ignore-certificate-errors
option as an argument to the executable. But in version 79, they stopped
doing so, and therefore we now have to specify that we want to accept
insecure certs. In these unit tests, the cert is insecure because it is
the one we generate to MITM the traffic.